std::hash implementation for all formula handles
